@JssDWt 2024-09-04
> BTW this 2x improvement in runtime is great! Do you know how big was this array of hints typically? I thought these hints were gathered on a per payment basis and hence there wasn't supposed to be many of them.

Thing is your own private channels are part of the hints. Being a LSP, the node has many private channels. In this case close to 2000.