@cdecker 2024-09-04
This conflicts with #7494 which essentially replaces the entire `channel_hint` handling, and encapsulates it in a separate file (and adds the `channel_hint_set` struct which is very similar to `channel_hint_map` here). This will likely need to be rebased on top of #7494