@rustyrussell 2024-10-16
Replaced by https://github.com/ElementsProject/lightning/pull/7726